requirements.txt
openai
pyyaml
jsonschema
python-dotenv